zipenhancer-standalone/requirements.txt
onnxruntime-gpu>=1.16.0
soundfile>=0.12.0
numpy>=1.21.0
scipy>=1.7.0
librosa>=0.10.0